refactor(lsp,mcp): extract reusable stdio server entry points
Move the fallow-lsp and fallow-mcp binary bodies into library crates
exposing run_stdio_server(), leaving each main.rs as a thin delegator.
Each entry builds its own multi-threaded tokio runtime (replacing
tokio::main) so a synchronous caller can drive it, and preserves the
version-probe, stdio wiring, and exit semantics of the standalone
binaries unchanged. This makes both server entries callable from a
future multicall fallow binary without duplicating the runtime setup.
